pyproject.toml
[project]
name = "smartscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agent-framework>=1.0.0b251120",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
